restructure project files, modify README, add .env.example file
