+++ .env.example
@@ -1,0 +1,12 @@
+# AWS variables to access data
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
+AWS_DEFAULT_REGION=
+AWS_S3_ENDPOINT=
+AWS_SESSION_TOKEN=
+
+# MLflow variables
+
+MLFLOW_TRACKING_URI=
+MLFLOW_TRACKING_USERNAME=
+MLFLOW_TRACKING_PASSWORD=
