--- _quarto.yml
@@ -1,23 +1,0 @@
-project:
-  type: website
-  render:
-    - index.qmd 
-
-website:
-  title: "Prévisions de données"
-  navbar:
-    left:
-      - href: index.qmd
-        text: Home
-      - about.qmd
-
-format:
-  html:
-    theme:
-      - cosmo
-      - brand
-    css: styles.css
-    toc: true
-
-
-
