+++ src/train/train_taxi.py
@@ -8,9 +8,9 @@ import mlflow
 import torch
 from dotenv import load_dotenv
 
-from data.data_loader import DataLoaderS3
-from data.data_preprocessing import DataConfig
-from data.plot_figures import plot_times_series
+from data_processing.data_loader import DataLoaderS3
+from data_processing.data_preprocessing import DataConfig
+from data_processing.plot_figures import plot_times_series
 from model.eval_model import error, eval_models
 from model.train_model import Trainer, TrainingConfig
