+++ tests/test_data_preprocessing.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import torch
 
-from data.data_preprocessing import (
+from data_processing.data_preprocessing import (
     DataConfig,
     create_time_series_window,
     get_normalization_metrics,
