+++ tests/test_train_model.py
@@ -1,9 +1,12 @@
 """Training tests."""
 
+import os
+
+import mlflow
 import pandas as pd
 import torch
 
-from data.data_preprocessing import DataConfig
+from data_processing.data_preprocessing import DataConfig
 from model.train_model import Trainer, TrainingConfig
 
 NB_MODELS_TRAINED = 2
@@ -11,6 +14,9 @@ NB_MODELS_TRAINED = 2
 
 def test_training() -> None:
     """Test the training with dummy data."""
+    mlflow_server = os.getenv("MLFLOW_TRACKING_URI")
+    mlflow.set_tracking_uri(mlflow_server)
+    mlflow.set_experiment("pytest dummy experiment")
     data_config = DataConfig(
         split_train=0.6,
         split_val=0.2,
@@ -24,7 +30,7 @@ def test_training() -> None:
     training_config = TrainingConfig(
         hidden_size=10,
         epochs=1,
-        batch_size=50,
+        batch_size=512,
         lr=1e-3,
         gammas=[1],
         max_norm=100.0,
@@ -43,7 +49,8 @@ def test_training() -> None:
 
     trainer = Trainer(dummy_data, device, data_config, training_config)
 
-    models = trainer.train_models()
+    with mlflow.start_run():
+        models = trainer.train_models()
 
     assert models is not None, "Les modèles n'ont pas été entraînés correctement."
     assert len(models) == NB_MODELS_TRAINED, "Il n'y a pas le bon nombre de modèles."
